+++ lib/src/features/chat/presentation/bloc/chat_event.dart
