+++ test/src/features/chat/presentation/pages/archieve_page_test.dart
@@ -6,10 +6,10 @@ import 'package:mocktail/mocktail.dart';
 import 'package:doormer/src/features/chat/domain/entities/chat_entity.dart';
 import 'package:doormer/src/features/chat/domain/repositories/chat_repository.dart';
 import 'package:doormer/src/features/chat/domain/usecases/archive_chat.dart';
-import 'package:doormer/src/features/chat/presentation/bloc/archive_bloc.dart';
+import 'package:doormer/src/features/chat/presentation/bloc/chat_bloc.dart';
 import 'package:doormer/src/features/chat/presentation/pages/archive_page.dart';
-import 'package:doormer/src/features/chat/presentation/bloc/archive_event.dart';
-import 'package:doormer/src/features/chat/presentation/bloc/archive_state.dart'
+import 'package:doormer/src/features/chat/presentation/bloc/chat_event.dart';
+import 'package:doormer/src/features/chat/presentation/bloc/chat_state.dart'
     as archive_state;
 import 'package:doormer/src/features/chat/presentation/widgets/chat_card.dart';
